Extracted ListViewModel
